vcs/*: move vcs/util.Rel function to new vcs/internal package.
