+++ vcs/gitcmd/repo.go
@@ -20,6 +20,7 @@ import (
 	"time"
 
 	"sourcegraph.com/sourcegraph/go-vcs/vcs"
+	"sourcegraph.com/sourcegraph/go-vcs/vcs/internal"
 	"sourcegraph.com/sourcegraph/go-vcs/vcs/util"
 	"sourcegraph.com/sqs/pbtypes"
 
@@ -848,7 +849,7 @@ type gitFSCmd struct {
 }
 
 func (fs *gitFSCmd) Open(name string) (vfs.ReadSeekCloser, error) {
-	name = util.Rel(name)
+	name = internal.Rel(name)
 	fs.repoEditLock.RLock()
 	defer fs.repoEditLock.RUnlock()
 	b, err := fs.readFileBytes(name)
@@ -887,7 +888,7 @@ func (fs *gitFSCmd) Lstat(path string) (os.FileInfo, error) {
 	fs.repoEditLock.RLock()
 	defer fs.repoEditLock.RUnlock()
 
-	path = filepath.Clean(util.Rel(path))
+	path = filepath.Clean(internal.Rel(path))
 
 	if path == "." {
 		// Special case root, which is not returned by `git ls-tree`.
@@ -932,7 +933,7 @@ func (fs *gitFSCmd) getModTimeFromGitLog(path string) (time.Time, error) {
 }
 
 func (fs *gitFSCmd) Stat(path string) (os.FileInfo, error) {
-	path = util.Rel(path)
+	path = internal.Rel(path)
 
 	fs.repoEditLock.RLock()
 	defer fs.repoEditLock.RUnlock()
@@ -961,7 +962,7 @@ func (fs *gitFSCmd) ReadDir(path string) ([]os.FileInfo, error) {
 	defer fs.repoEditLock.RUnlock()
 	// Trailing slash is necessary to ls-tree under the dir (not just
 	// to list the dir's tree entry in its parent dir).
-	return fs.lsTree(filepath.Clean(util.Rel(path)) + "/")
+	return fs.lsTree(filepath.Clean(internal.Rel(path)) + "/")
 }
 
 func (fs *gitFSCmd) lsTree(path string) ([]os.FileInfo, error) {
